+++ accessws/aw_order.h
@@ -8,9 +8,8 @@
 
 int init_order(void);
 
-int order_subscribe(uint32_t user_id, nw_ses *ses, const char *market);
-int order_unsubscribe(uint32_t user_id, nw_ses *ses);
-int order_on_update(uint32_t user_id, int event, json_t *order);
+int order_subscribe(uint64_t user_id, nw_ses *ses, const char *market);
+int order_unsubscribe(uint64_t user_id, nw_ses *ses);
+int order_on_update(uint64_t user_id, int event, json_t *order);
 
 # endif
-
